RENAME: metabuild_functions --> stats
